requirements.txt
fastapi
uvicorn
requests
apscheduler
google-generativeai